Resolve code review suggestions
- Remove Chunk force inline
- Remove ChunkTile force inline
- Remove RenderController unused member
- MouseSelection only fetch cursor sprite when necessary
- Fix world clear not clearing tex coord ids
- Remove unnecessary unbind in opengl object destructor
- Remove unncessary state changes/bindings in imgui renderer
- Remove proto_renderer unnecessary conversion double -> float
- Use SafeCast/LossyCast when casting numbers
- Fix spritemap_genreator multiply then widen, instead of widen then multiply
- Reduce computations when setting pixel in spritemap generator
- Fix TRenderBuffer implicit conversion

TileRenderer
- Remove unnecessary depth buffer clear
- Remove unnecessary member
- Remove unused method
- Clean up confusing calculations
- Fix incorrect capacity check
- Fix incorrect calculation of tiles to prepare

DVector
- Fix incorrect reserve when moving elements up/down
- Remove emplace for now (current emplace offers no way to modify index -1)
- Fix missing move in move constructor
- Fix incorrect copy, move behavior in DVector test's TestClass

- Document potentially unexpected behavior in methods
- Document in tests reason for attaching sprite with tex coord id
- Remove incorrect comments

+++ src/render/trender_buffer.cpp
@@ -60,7 +60,7 @@ void render::TRenderBuffer::GlHandleBufferResize() {
         return;
 
     gResizeVertexBuffers_ = false;
-    eCapacity_            = LossyCast<decltype(eCapacity_)>(queuedECapacity_ * kResizeECapacityMultiplier);
+    eCapacity_ = LossyCast<decltype(eCapacity_)>(SafeCast<float>(queuedECapacity_) * kResizeECapacityMultiplier);
 
     baseVb_.Reserve(nullptr, eCapacity_ * kBaseBytesPerElement, false);
